fix the bug that deleting wrong record

+++ controllers/likeController.js
@@ -4,9 +4,10 @@ const helpers = require('../_helpers')
 const likeController = {
   postUnlike: async (req, res) => {
     try {
-      const targetLike = await Like.findOne({ raw: true, nest: true, where: { TweetId: Number(req.params.id) } })
-      if (targetLike.UserId === helpers.getUser(req).id) {
-        await Like.destroy({ where: { TweetId: Number(req.params.id) } })
+      const targetLikes = await Like.findAll({ raw: true, nest: true, where: { TweetId: Number(req.params.id) } })
+      const isUserLiked = targetLikes.some(data => data.UserId === helpers.getUser(req).id)
+      if (isUserLiked) {
+        await Like.destroy({ where: { TweetId: Number(req.params.id), UserId: Number(helpers.getUser(req).id) } })
         return res.status(200).json({ status: 'success', message: '成功取消喜歡' })
       } else {
         // filter out the scenario that trying to alter others following list
